Improve homomorphic encryption module and stress test reporting
Update homomorphic encryption module by marking unused NTT functions with __attribute__((unused)), correcting potential buffer underflow error message formatting, and adjusting printf format specifiers for large integers (e.g., operations performed, total operation time, creation timestamp) in context and ciphertext info printing. Also, a variable `peak_memory` in the stress test function was commented out as it was unused, and the progression reporting within the stress test was updated to use the correct format specifier for `total_operations`. Finally, the stress test success reporting was updated to use the correct format specifier for `total_operations`.

+++ src/main.c
@@ -1635,7 +1635,7 @@ bool he_stress_test_100m_operations_wrapper(void) {
         if (result->test_success) {
             printf("  ✅ STRESS TEST RÉUSSI!\n");
             printf("  📊 Opérations: %llu en %.3f secondes\n", 
-                   result->total_operations,
+                   (unsigned long long)result->total_operations,
                    (double)result->total_time_ns / 1000000000.0);
             printf("  ⚡ Débit: %.0f opérations/seconde\n", 
                    result->operations_per_second);
